+++ train.py
@@ -82,7 +82,7 @@ if __name__ == '__main__':
     # Model
     logger.info('Building model...')
     model = FullDPM(n_layers=args.layers,additional_layers=args.add_layers).to(args.device)
-    model.double()
+    # model.double()
     logger.info('Number of parameters: %d' % count_parameters(model))
     logger.info('Number of EGCL layers: %d' % args.layers)
     logger.info('Number of additional layers: %d' % args.add_layers)
